// frontend/src/interfaces/servicePlanBooking.ts
export interface ServicePlanBooking {
  pet_species: string;
  pet_breed: string;
  pet_size: string;
  pet_gender: string;
  pet_age: string;
  aggressiveness: string;
  user_name: string;
  email: string;
  phoneNo: string;
  address: string;
  city: string;
  state: string;
  notes?: string;
  booking_date: string;
  booking_time: string;
}